% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capfunding2526.R
\docType{data}
\name{capfunding2526_schools}
\alias{capfunding2526_schools}
\alias{capfunding2526_rb}
\alias{capfunding2526_nmss}
\alias{capfunding2526_spi}
\title{School Capital Funding Allocations, 2025-26}
\format{
Each is a tibble with one record per institution, showing the funding
amount and various funding factors associated with it.
}
\source{
\url{https://www.gov.uk/guidance/school-capital-funding}
}
\usage{
capfunding2526_schools

capfunding2526_rb

capfunding2526_nmss

capfunding2526_spi
}
\description{
Published dataset of school capital funding. See the source for full details.
There are 4 datasets:
\itemize{
\item \code{capfunding2526_schools} - school allocations (devolved formula capital, DFC)
\item \code{capfunding2526_rb} - responsible bodies (school condition allocation, SCA)
\item \code{capfunding2526_nmss} - non-maintained special schools (DFC & SCA)
\item \code{capfunding2526_spi} - special post-16 institutions (DFC & SCA)
}
}
\keyword{datasets}
