% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census21_pop.R
\docType{data}
\name{census21_pop}
\alias{census21_pop}
\title{Census 2021 Population Estimates}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 214032 rows and 4 columns.
}
\source{
\url{https://www.ons.gov.uk/datasets/create}
}
\usage{
census21_pop
}
\description{
A download of usual residents by age group at LSOA level from the 2021 census.
}
\keyword{datasets}
