% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif2425.R
\docType{data}
\name{cif2425}
\alias{cif2425}
\title{Condition Improvement Fund Outcomes, 2024-25}
\format{
A tibble giving the school name and URN, and the project name
}
\source{
\url{https://www.gov.uk/government/publications/condition-improvement-fund-2024-to-2025-outcome}
}
\usage{
cif2425
}
\description{
The published list of approved (and pending) projects funded by CIF.
}
\keyword{datasets}
