% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gias.R
\docType{data}
\name{gias}
\alias{gias}
\title{Get Information About Schools}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27153 rows and 16 columns.
}
\source{
\url{https://get-information-schools.service.gov.uk/Downloads}
}
\usage{
gias
}
\description{
An extract of all open state-funded schools, taken from the Get Information
About Schools Service on 2025-05-24.
}
\keyword{datasets}
