% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iod2019.R
\docType{data}
\name{iod2019_eng}
\alias{iod2019_eng}
\title{English Indices of Deprivation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32844 rows and 20 columns.
}
\source{
\url{https://www.gov.uk/government/statistics/english-indices-of-deprivation-2019}
}
\usage{
iod2019_eng
}
\description{
Ranks and deciles by deprivation domain for each LSOA in England, 2019.
}
\keyword{datasets}
