% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iod2019.R
\docType{data}
\name{iod2019_wal}
\alias{iod2019_wal}
\title{Welsh Indices of Deprivation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1909 rows and 20 columns.
}
\source{
\url{https://www.gov.wales/welsh-index-multiple-deprivation-full-index-update-ranks-2019}
}
\usage{
iod2019_wal
}
\description{
Ranks and deciles by deprivation domain for each LSOA in Wales, 2019.
}
\keyword{datasets}
