% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsoa2011.R
\docType{data}
\name{lsoa2011_centroids}
\alias{lsoa2011_centroids}
\title{Lower-Layer Super Output Area Centroids, 2011}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34753 rows and 2 columns.
}
\source{
\url{https://geoportal.statistics.gov.uk/search?q=BDY_LSOA\%20DEC_2011}
}
\usage{
lsoa2011_centroids
}
\description{
Boundaries and centroids for LSOAs from the 2011 census. 2011 is used, not
2021, since the deprivation dataset in this repo uses the 2011 boundaries.
Boundaries are super-generalised and clipped.
}
\keyword{datasets}
