% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\docType{data}
\name{mps}
\alias{mps}
\title{Members of Parliament}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 650 rows and 8 columns.
}
\source{
\url{https://members.parliament.uk/members/commons}
}
\usage{
mps
}
\description{
A data frame of the UK members of parliament, their constituency and party.
}
\keyword{datasets}
