% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcon2024.R
\docType{data}
\name{pcon2024}
\alias{pcon2024}
\title{Parliamentary Constituency Boundaries, July 2024}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 650 rows and 5 columns.
}
\source{
\url{https://geoportal.statistics.gov.uk/search?q=BDY_ELE}
}
\usage{
pcon2024
}
\description{
Boundaries and centroids for Westminster Constituencies, as at the July 2024
general election. Boundaries are ultra-generalised and clipped.
}
\keyword{datasets}
