% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsoa2011.R
\docType{data}
\name{lsoa2011}
\alias{lsoa2011}
\alias{lsoa2011_centroids}
\title{Lower-Layer Super Output Areas & Centroids, 2011}
\format{
Both are spatial datasets requiring the \code{sf} package to be loaded to
print nicely in the console.
}
\source{
\url{https://geoportal.statistics.gov.uk/search?q=BDY_LSOA\%20DEC_2011}
}
\usage{
lsoa2011

lsoa2011_centroids
}
\description{
Boundaries and centroids for LSOAs from the 2011 census. 2011 is used, not
2021, since the deprivation dataset in this repo uses the 2011 boundaries.
Boundaries are super-generalised and clipped.
}
\keyword{datasets}
