% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcon2024.R
\docType{data}
\name{pcon2024}
\alias{pcon2024}
\alias{pcon2024_centroids}
\title{Parliamentary Constituencies & Centroids, July 2024}
\format{
Both are spatial datasets requiring the \code{sf} package to be loaded to
print nicely in the console.
}
\source{
\url{https://geoportal.statistics.gov.uk/search?q=BDY_ELE}
}
\usage{
pcon2024

pcon2024_centroids
}
\description{
Boundaries and centroids for Westminster Constituencies, as at the July 2024
general election. Boundaries are ultra-generalised and clipped.
}
\keyword{datasets}
